AWS_Lambda/requirements.txt
paho-mqtt==1.5.1
google-cloud-firestore==2.9.1